Many *.c files: hundreds of small changes to eliminate compiler warning messages on various types of systems.

+++ trunk/src/mbio/mb_io.h
@@ -427,7 +427,7 @@ struct mb_io_struct
 	void	**regarray_ptr;
 	void	**regarray_oldptr;
 	int	*regarray_type;
-	int	*regarray_size;
+	size_t	*regarray_size;
 
 	/* variables for saving information */
 	char	save_label[12];
